enos_vault_start: support configuring the retry_join block in storage
In order to support forcing `retry_join` to resolve using IPV6 we need
to include the address type in the `auto_join` stanza. Historically we
have handled the `retry_join` stanza in the `enos_vault_start` resource
and did not have a way to define the `retry_join` block.

This does just that by supporting `retry_join` in the storage schema by
using a dynamic pseudo type for `storage` in the config.

I've tested the resource with no `retry_join` defined and added a new
variant to our test scenarios to test with it set or null. All work just
fine.
